+++ tools/libexec/perl/TTP/IFindable.pm
@@ -161,9 +161,9 @@ sub _find_single {
 		push( @results, $fname );
 	}
 	foreach my $candidate ( @results ){
-		print STDERR __PACKAGE__."::_find_single() testing '$candidate'".EOL if $ENV{TTP_DEBUG};
+		msgDebug( __PACKAGE__."::_find_single() testing '$candidate'" );
 		if( -r $candidate ){
-			print STDERR __PACKAGE__."::_find_single() found candidate '$candidate'".EOL if $ENV{TTP_DEBUG};
+			msgDebug( __PACKAGE__."::_find_single() found candidate '$candidate'" );
 			push( @{$self->{_ifindable}{candidates}}, $candidate );
 			my $accepted = true;
 			if( $self->does( 'TTP::IAcceptable' ) && $opts->{acceptable} ){
@@ -174,9 +174,9 @@ sub _find_single {
 			if( $accepted ){
 				push( @{$self->{_ifindable}{accepted}}, $candidate );
 				$self->{_ifindable}{end} = true unless $self->{_ifindable}{wantsAll};
-				print STDERR __PACKAGE__."::_find_inpath() candidate='$candidate' is accepted".EOL if $ENV{TTP_DEBUG};
+				msgDebug( __PACKAGE__."::_find_inpath() candidate='$candidate' is accepted" );
 			} else {
-				print STDERR __PACKAGE__."::_find_inpath() candidate='$candidate' is refused".EOL if $ENV{TTP_DEBUG};
+				msgDebug( __PACKAGE__."::_find_inpath() candidate='$candidate' is refused" );
 			}
 		}
 	}