+++ ChangeLog.md
@@ -23,6 +23,8 @@
     - Change the default default values of ttp.pl alerts from true to false, keeping default enabled to true
     - Define the default default values of telemetry.pl publish to false, having default enabled to true
     - dbms.pl backup and dbms.pl restore have and honor '--report' option
+    - Introduce TTP::EP->bootstrapped()
+    - Introduce msgDebug()
 
 ### 4.7.2